requirements.txt
# Core ML Framework
torch>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Data Processing
pandas>=2.0.0
polars>=0.19.0

# Time-Series
pytorch-forecasting>=1.0.0

# Tree-based Models
xgboost>=2.0.0
lightgbm>=4.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Evaluation Metrics
scikit-learn>=1.3.0
scikit-survival>=0.22.0
lifelines>=0.27.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Interpretability
shap>=0.43.0

# Experiment Tracking
mlflow>=2.8.0
wandb>=0.15.0

# Utilities
tqdm>=4.66.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0  # For RxNorm API calls

# OCR and Document Processing
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0
easyocr>=1.7.0  # Optional: More accurate but slower
python-dateutil>=2.8.2  # Date parsing for temporal data

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.10.0
flake8>=6.1.0
mypy>=1.7.0

